update website common utils
